neighbors: split dispatch tests by backend interface
Move Torch and JAX dispatch report coverage into their bindings-specific
test directories. Keep the shared base dispatch test focused on backend-neutral
finalization and cell-list helper behavior, removing the parametrized report
axis.
